Adding support for STSAssumeRoleCredentialProvider
